requirements.txt
streamlit
flask
groq
python-dotenv
pandas
requests
PyPDF2
